requirements.txt
chess
pyautogui
opencv-python

# TensorFlow with GPU support (version can be adjusted)
tensorflow==2.12.0

# Common data science libraries
numpy
pandas
scikit-learn
matplotlib
seaborn

# Additional useful libraries
scipy
h5py
pillow